Deduplicate access log and TLS cipher suite validation logic
Had to do both because of package import issues

Removed ability to configure cipher suites with leading/trailing
whitespace, that is now a config validation error.

Tried to keep all exported vars/methods in pkg/config but used the
implementations in apis/projectcontour/v1alpha1

Kept some simple tests in the pkg/config tests but moved the more
detailed things over to apis/projectcontour/v1alpha1

Also ensures recent changes from #4610 are made available in config CRD
