+++ apis/projectcontour/v1alpha1/accesslog.go
@@ -42,6 +42,7 @@ var DefaultAccessLogJSONFields = AccessLogJSONFields([]string{
 	"upstream_service_time",
 	"user_agent",
 	"x_forwarded_for",
+	"grpc_status",
 })
 
 // DefaultAccessLogType is the default access log format.
@@ -72,6 +73,9 @@ var envoySimpleOperators = map[string]struct{}{
 	"CONNECTION_TERMINATION_DETAILS":                {},
 	"DOWNSTREAM_DIRECT_REMOTE_ADDRESS":              {},
 	"DOWNSTREAM_DIRECT_REMOTE_ADDRESS_WITHOUT_PORT": {},
+	"DOWNSTREAM_DIRECT_REMOTE_PORT":                 {},
+	"DOWNSTREAM_HEADER_BYTES_RECEIVED":              {},
+	"DOWNSTREAM_HEADER_BYTES_SENT":                  {},
 	"DOWNSTREAM_LOCAL_ADDRESS":                      {},
 	"DOWNSTREAM_LOCAL_ADDRESS_WITHOUT_PORT":         {},
 	"DOWNSTREAM_LOCAL_PORT":                         {},
@@ -88,32 +92,53 @@ var envoySimpleOperators = map[string]struct{}{
 	"DOWNSTREAM_PEER_URI_SAN":                       {},
 	"DOWNSTREAM_REMOTE_ADDRESS":                     {},
 	"DOWNSTREAM_REMOTE_ADDRESS_WITHOUT_PORT":        {},
+	"DOWNSTREAM_REMOTE_PORT":                        {},
 	"DOWNSTREAM_TLS_CIPHER":                         {},
 	"DOWNSTREAM_TLS_SESSION_ID":                     {},
 	"DOWNSTREAM_TLS_VERSION":                        {},
+	"DOWNSTREAM_WIRE_BYTES_RECEIVED":                {},
+	"DOWNSTREAM_WIRE_BYTES_SENT":                    {},
 	"DURATION":                                      {},
+	"FILTER_CHAIN_NAME":                             {},
 	"GRPC_STATUS":                                   {},
+	"GRPC_STATUS_NUMBER":                            {},
 	"HOSTNAME":                                      {},
 	"LOCAL_REPLY_BODY":                              {},
 	"PROTOCOL":                                      {},
+	"REQUEST_HEADERS_BYTES":                         {},
 	"REQUESTED_SERVER_NAME":                         {},
 	"REQUEST_DURATION":                              {},
+	"REQUEST_TX_DURATION":                           {},
 	"RESPONSE_CODE":                                 {},
 	"RESPONSE_CODE_DETAILS":                         {},
 	"RESPONSE_DURATION":                             {},
 	"RESPONSE_FLAGS":                                {},
+	"RESPONSE_HEADERS_BYTES":                        {},
+	"RESPONSE_TRAILERS_BYTES":                       {},
 	"RESPONSE_TX_DURATION":                          {},
 	"ROUTE_NAME":                                    {},
 	"START_TIME":                                    {},
 	"UPSTREAM_CLUSTER":                              {},
+	"UPSTREAM_HEADER_BYTES_RECEIVED":                {},
+	"UPSTREAM_HEADER_BYTES_SENT":                    {},
 	"UPSTREAM_HOST":                                 {},
 	"UPSTREAM_LOCAL_ADDRESS":                        {},
+	"UPSTREAM_LOCAL_ADDRESS_WITHOUT_PORT":           {},
+	"UPSTREAM_LOCAL_PORT":                           {},
+	"UPSTREAM_REMOTE_ADDRESS":                       {},
+	"UPSTREAM_REMOTE_ADDRESS_WITHOUT_PORT":          {},
+	"UPSTREAM_REMOTE_PORT":                          {},
+	"UPSTREAM_REQUEST_ATTEMPT_COUNT":                {},
 	"UPSTREAM_TRANSPORT_FAILURE_REASON":             {},
+	"UPSTREAM_WIRE_BYTES_RECEIVED":                  {},
+	"UPSTREAM_WIRE_BYTES_SENT":                      {},
+	"VIRTUAL_CLUSTER_NAME":                          {},
 }
 
 // envoyComplexOperators is the list of known Envoy log template keywords that require
 // arguments.
 var envoyComplexOperators = map[string]struct{}{
+	"ENVIRONMENT":       {},
 	"REQ":               {},
 	"RESP":              {},
 	"START_TIME":        {},
@@ -155,7 +180,7 @@ func (a AccessLogJSONFields) Validate() error {
 			continue
 		}
 
-		err := parseAccessLogFormat(val)
+		err := parseAccessLogFormatString(val)
 		if err != nil {
 			return fmt.Errorf("invalid JSON field: %s", err)
 		}
@@ -214,16 +239,18 @@ const (
 	LogLevelDisabled AccessLogLevel = "disabled"
 )
 
-func validateAccessLogFormatString(format string) error {
+type AccessLogFormatString string
+
+func (s AccessLogFormatString) Validate() error {
 	// Empty format means use default format, defined by Envoy.
-	if format == "" {
+	if s == "" {
 		return nil
 	}
-	err := parseAccessLogFormat(format)
+	err := parseAccessLogFormatString(string(s))
 	if err != nil {
 		return fmt.Errorf("invalid access log format: %s", err)
 	}
-	if !strings.HasSuffix(format, "\n") {
+	if !strings.HasSuffix(string(s), "\n") {
 		return fmt.Errorf("invalid access log format: must end in newline")
 	}
 	return nil
@@ -241,11 +268,10 @@ func validateAccessLogFormatString(format string) error {
 //   4. Truncation length: ":3"
 var commandOperatorRegexp = regexp.MustCompile(`%(([A-Z_]+)(\([^)]+\)(:[0-9]+)?)?%)?`)
 
-func parseAccessLogFormat(format string) error {
-
+func parseAccessLogFormatString(format string) error {
 	// FindAllStringSubmatch will always return a slice with matches where every slice is a slice
 	// of submatches with length of 5 (number of capture groups + 1).
-	tokens := commandOperatorRegexp.FindAllStringSubmatch(format, -1)
+	tokens := commandOperatorRegexp.FindAllStringSubmatch(string(format), -1)
 	if len(tokens) == 0 {
 		return nil
 	}
@@ -262,7 +288,7 @@ func parseAccessLogFormat(format string) error {
 			return fmt.Errorf("invalid Envoy format: %s, invalid Envoy operator: %s", f, op)
 		}
 
-		if (op == "REQ" || op == "RESP" || op == "TRAILER" || op == "REQ_WITHOUT_QUERY") && f[3] == "" {
+		if (op == "REQ" || op == "RESP" || op == "TRAILER" || op == "REQ_WITHOUT_QUERY" || op == "ENVIRONMENT") && f[3] == "" {
 			return fmt.Errorf("invalid Envoy format: %s, arguments required for operator: %s", f, op)
 		}
