+++ examples/render/contour-deployment.yaml
@@ -402,19 +402,19 @@ spec:
                               \n Contour's default list is:   - \"[ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]\"
                               \  - \"[ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]\"
                               \  - \"ECDHE-ECDSA-AES256-GCM-SHA384\"   - \"ECDHE-RSA-AES256-GCM-SHA384\"
-                              \n Contour recommends leaving this undefined unless
-                              you are sure you must. \n See: https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/transport_sockets/tls/v3/common.proto#extensions-transport-sockets-tls-v3-tlsparameters
+                              \n Ciphers provided are validated against the following
+                              list:   - \"[ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]\"
+                              \  - \"[ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]\"
+                              \  - \"ECDHE-ECDSA-AES128-GCM-SHA256\"   - \"ECDHE-RSA-AES128-GCM-SHA256\"
+                              \  - \"ECDHE-ECDSA-AES128-SHA\"   - \"ECDHE-RSA-AES128-SHA\"
+                              \  - \"AES128-GCM-SHA256\"   - \"AES128-SHA\"   - \"ECDHE-ECDSA-AES256-GCM-SHA384\"
+                              \  - \"ECDHE-RSA-AES256-GCM-SHA384\"   - \"ECDHE-ECDSA-AES256-SHA\"
+                              \  - \"ECDHE-RSA-AES256-SHA\"   - \"AES256-GCM-SHA384\"
+                              \  - \"AES256-SHA\" \n Contour recommends leaving this
+                              undefined unless you are sure you must. \n See: https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/transport_sockets/tls/v3/common.proto#extensions-transport-sockets-tls-v3-tlsparameters
                               Note: This list is a superset of what is valid for stock
                               Envoy builds and those using BoringSSL FIPS."
                             items:
-                              description: "TLSCipherType is a string alias for the
-                                TLS ciphers supported by Envoy. \n Values: \n * [ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]
-                                * [ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]
-                                * ECDHE-ECDSA-AES128-GCM-SHA256 * ECDHE-RSA-AES128-GCM-SHA256
-                                * ECDHE-ECDSA-AES128-SHA * ECDHE-RSA-AES128-SHA *
-                                AES128-GCM-SHA256 * AES128-SHA * ECDHE-ECDSA-AES256-GCM-SHA384
-                                * ECDHE-RSA-AES256-GCM-SHA384 * ECDHE-ECDSA-AES256-SHA
-                                * ECDHE-RSA-AES256-SHA * AES256-GCM-SHA384 * AES256-SHA"
                               type: string
                             type: array
                           minimumProtocolVersion:
@@ -1405,20 +1405,20 @@ spec:
                                   \  - \"[ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]\"
                                   \  - \"[ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]\"
                                   \  - \"ECDHE-ECDSA-AES256-GCM-SHA384\"   - \"ECDHE-RSA-AES256-GCM-SHA384\"
-                                  \n Contour recommends leaving this undefined unless
+                                  \n Ciphers provided are validated against the following
+                                  list:   - \"[ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]\"
+                                  \  - \"[ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]\"
+                                  \  - \"ECDHE-ECDSA-AES128-GCM-SHA256\"   - \"ECDHE-RSA-AES128-GCM-SHA256\"
+                                  \  - \"ECDHE-ECDSA-AES128-SHA\"   - \"ECDHE-RSA-AES128-SHA\"
+                                  \  - \"AES128-GCM-SHA256\"   - \"AES128-SHA\"   -
+                                  \"ECDHE-ECDSA-AES256-GCM-SHA384\"   - \"ECDHE-RSA-AES256-GCM-SHA384\"
+                                  \  - \"ECDHE-ECDSA-AES256-SHA\"   - \"ECDHE-RSA-AES256-SHA\"
+                                  \  - \"AES256-GCM-SHA384\"   - \"AES256-SHA\" \n
+                                  Contour recommends leaving this undefined unless
                                   you are sure you must. \n See: https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/transport_sockets/tls/v3/common.proto#extensions-transport-sockets-tls-v3-tlsparameters
                                   Note: This list is a superset of what is valid for
                                   stock Envoy builds and those using BoringSSL FIPS."
                                 items:
-                                  description: "TLSCipherType is a string alias for
-                                    the TLS ciphers supported by Envoy. \n Values:
-                                    \n * [ECDHE-ECDSA-AES128-GCM-SHA256|ECDHE-ECDSA-CHACHA20-POLY1305]
-                                    * [ECDHE-RSA-AES128-GCM-SHA256|ECDHE-RSA-CHACHA20-POLY1305]
-                                    * ECDHE-ECDSA-AES128-GCM-SHA256 * ECDHE-RSA-AES128-GCM-SHA256
-                                    * ECDHE-ECDSA-AES128-SHA * ECDHE-RSA-AES128-SHA
-                                    * AES128-GCM-SHA256 * AES128-SHA * ECDHE-ECDSA-AES256-GCM-SHA384
-                                    * ECDHE-RSA-AES256-GCM-SHA384 * ECDHE-ECDSA-AES256-SHA
-                                    * ECDHE-RSA-AES256-SHA * AES256-GCM-SHA384 * AES256-SHA"
                                   type: string
                                 type: array
                               minimumProtocolVersion:
